@@ -1,12 +1,11 @@
-﻿namespace PeopleWhoCanCode.DatabaseVersioning.Models
-{
-    public class ConnectionString : IConnectionString
-    {
-        public string Value { get; }
+﻿namespace PeopleWhoCanCode.DatabaseVersioning.Models;
 
-        public ConnectionString(string value)
-        {
-            Value = value;
-        }
+public class ConnectionString : IConnectionString
+{
+    public string Value { get; }
+
+    public ConnectionString(string value)
+    {
+        Value = value;
     }
 }
