@@ -1,14 +1,11 @@
-﻿using System;
+﻿namespace PeopleWhoCanCode.DatabaseVersioning.Exceptions;
 
-namespace PeopleWhoCanCode.DatabaseVersioning.Exceptions
+public class InvalidChangeScriptNumberException : Exception
 {
-    public class InvalidChangeScriptNumberException : Exception
-    {
-        public string InvalidNumber { get; }
+    public string InvalidNumber { get; }
 
-        public InvalidChangeScriptNumberException(string invalidNumber)
-        {
-            InvalidNumber = invalidNumber;
-        }
+    public InvalidChangeScriptNumberException(string invalidNumber)
+    {
+        InvalidNumber = invalidNumber;
     }
 }
